+++ changelog.md
@@ -1,3 +1,10 @@
+v0.2.1
+------
+
+- Added track width calculator (not operational yet).
+- NinjaCalc app now has proper icon.
+- Nav menu teaser is now bigger and with "justified text" button.
+
 v0.2.0
 ------